refactor(svm): move the instruction catalog out of system_config
instruction_catalog lives in svm_catalog.rs. IxIdl.derived and Dispatch
are gone; dispatch is the discriminator bytes.
