+++ packages/cli/src/config_parsing/mod.rs
@@ -8,6 +8,7 @@ pub mod field_types;
 pub mod human_config;
 pub mod hypersync_endpoints;
 pub mod public_config;
+pub mod svm_catalog;
 pub mod svm_idl;
 pub mod system_config;
 pub mod validation;
